Fix TypeScript error: Cast PDF buffer to Uint8Array

+++ src/app/api/session/[matchId]/generate-pdf/route.ts
@@ -240,7 +240,7 @@ export async function POST(
         })
         
         // PDF를 직접 반환
-        return new NextResponse(pdfBuffer, {
+        return new NextResponse(pdfBuffer as Uint8Array, {
           headers: {
             'Content-Type': 'application/pdf',
             'Content-Disposition': `attachment; filename="${matchId}.pdf"`
